qcacld-3.0: Update measurement request max array index
For radio measurement report, we can send maximum of
MAX_MGMT_MPDU_LEN (2304) / sizeof(tDot11fIEMeasurementReport)
reports which is 8. But currently only one instance of struct
tDot11fIEMeasurementReport is allocated and malloc for remaining
number of reports dynamically.

Instead allocate the max reports on the structure and use it.

CRs-Fixed: 3692872
Change-Id: I954d4fff1181cfa1e41b1321ba53c9fc7a35ae42